General Improvements

+++ include/main.hpp
@@ -10,6 +10,7 @@
 
 #include "common_package/commands.hpp"
 #include "common_package/common_node.hpp"
+#include "common_package/node_names.hpp"
 #include "common_package/topic_names.hpp"
 #include "rclcpp/rclcpp.hpp"
 #include "structs.hpp"
@@ -47,10 +48,10 @@ class WaypointNode : public common_lib::CommonNode {
 
     Position pos;  //!< Current position
 
-    static constexpr uint16_t max_position_msg_time_difference_ms =
+    static constexpr uint32_t max_position_msg_time_difference_ms =
         100;  //!< Maximum age of a position message from FCC bridge
 
-    static constexpr uint16_t max_progress_msg_time_difference_ms =
+    static constexpr uint32_t max_progress_msg_time_difference_ms =
         500;  //!< Maximum age of a progress message from FCC bridge
 
     float mission_progress = 0.0;  //!< Current mission progress
